Add isolated build tests to all packages except editor and dashboard
