+++ packages/map/src/test/CdcMap.test.jsx
@@ -2,7 +2,7 @@ import path from 'path'
 import { testStandaloneBuild } from '@cdc/core/helpers/tests/testStandaloneBuild.test.ts'
 import { describe, it, expect } from 'vitest'
 
-describe('Chart', () => {
+describe('Map', () => {
   it('Can be built in isolation', async () => {
     const pkgDir = path.join(__dirname, '..')
     const result = testStandaloneBuild(pkgDir)
